+++ K6/api/building-blocks/dialogporten/serviceowner/get-dialogs.js
@@ -60,217 +60,6 @@ export function GetDialog(
 }
 
 /**
- * Function to get dialog activities
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog to get activities for
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogActivities(
-    serviceOwnerApiClient,
-    dialogId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogActivities(
-        dialogId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogActivities - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get a dialog activity by id
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog the activity belongs to
- * param {string} activityId - id of the activity to get
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogActivity(
-    serviceOwnerApiClient,
-    dialogId,
-    activityId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogActivity(
-        dialogId,
-        activityId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogActivity - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get serviceowner labels
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog to get labels for
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetServiceOwnerLabels(
-    serviceOwnerApiClient,
-    dialogId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetServiceOwnerLabels(
-        dialogId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetServiceOwnerLabels - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get dialog seen log
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog to get seen log for
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogSeenLog(
-    serviceOwnerApiClient,
-    dialogId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogSeenLog(
-        dialogId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogSeenLog - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get dialog seen log entry
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog the seen log entry belongs to
- * param {string} seenLogEntryId - id of the seen log entry to get
- * @param {string} label - label for the request
- */
-export function GetDialogSeenLogEntry(
-    serviceOwnerApiClient,
-    dialogId,
-    seenLogEntryId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogSeenLogEntry(
-        dialogId,
-        seenLogEntryId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogSeenLogEntry - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get dialog tranmissions
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog to get transmissions for
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogTransmissions(
-    serviceOwnerApiClient,
-    dialogId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogTransmissions(
-        dialogId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogTransmissions - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get a dialog transmission by id
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} dialogId - id of the dialog the transmission belongs to
- * param {string} transmissionId - id of the transmission to get
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogTransmission(
-    serviceOwnerApiClient,
-    dialogId,
-    transmissionId,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogTransmission(
-        dialogId,
-        transmissionId,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogTransmission - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
  * Function to get enduser context
  * @param {ServiceOwnerApiClient} serviceOwnerApiClient
  * @param {string} queryParams - query parameters for the request
@@ -289,35 +78,6 @@ export function GetEndUserContext(
 
     const success = check(res, {
         "GetEndUserContext - status code MUST be 200": (res) => res.status == 200,
-    });
-
-    if (!success) {
-        console.log(res.status);
-        console.log(res.body);
-    }
-
-    return res.body;
-}
-
-/**
- * Function to get dialogLookup
- * @param {ServiceOwnerApiClient} serviceOwnerApiClient
- * @param {string} queryParams - query parameters for the request
- * @param {string} label - label for the request
- * @return response body of the request
- */
-export function GetDialogLookup(
-    serviceOwnerApiClient,
-    queryParams,
-    labels = null,
-) {
-    const res = serviceOwnerApiClient.GetDialogLookup(
-        queryParams,
-        labels,
-    );
-
-    const success = check(res, {
-        "GetDialogLookup - status code MUST be 200": (res) => res.status == 200,
     });
 
     if (!success) {
