+++ K6/api/building-blocks/dialogporten/serviceowner/index.js
@@ -1,15 +1,8 @@
 export { GetDialogsQueriesNotificationCondition } from "./get-dialogs-queries-notification-condition.js";
 export { CreateDialog, CreateTransmission, CreateActivity } from "./create-dialog.js";
-export {
-    GetDialogs,
-    GetDialog,
-    GetDialogActivities,
-    GetDialogActivity,
-    GetServiceOwnerLabels,
-    GetDialogSeenLog,
-    GetDialogSeenLogEntry,
-    GetDialogTransmissions,
-    GetDialogTransmission,
-    GetEndUserContext,
-    GetDialogLookup,
-} from "./get-dialogs.js";
+export { GetDialogs, GetDialog, GetEndUserContext } from "./get-dialogs.js";
+export { GetDialogActivities, GetDialogActivity } from "./get-dialog-activities.js";
+export { GetDialogSeenLogs, GetDialogSeenLog } from "./get-dialog-seen-logs.js";
+export { GetDialogTransmissions, GetDialogTransmission } from "./get-dialog-transmissions.js";
+export { GetServiceOwnerLabels } from "./get-service-owner-labels.js";
+export { GetDialogLookup } from "./get-dialog-lookup.js";
